+++ ntex-bytes/src/string.rs
@@ -17,7 +17,7 @@ impl ByteString {
     /// Get a str slice.
     #[inline]
     pub fn as_str(&self) -> &str {
-        &*self
+        self
     }
 
     /// Get a reference to the underlying bytes.
@@ -140,6 +140,23 @@ impl ByteString {
         self.0.trimdown()
     }
 
+    /// Clears the buffer, removing all data.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use ntex_bytes::ByteString;
+    ///
+    /// let mut a = ByteString::from("hello world");
+    /// a.clear();
+    ///
+    /// assert!(a.is_empty());
+    /// ```
+    #[inline]
+    pub fn clear(&mut self) {
+        self.0.clear()
+    }
+
     /// Creates a new `ByteString` from a Bytes.
     ///
     /// # Safety
@@ -168,7 +185,7 @@ impl<T: AsRef<str>> PartialEq<T> for ByteString {
 impl AsRef<str> for ByteString {
     #[inline]
     fn as_ref(&self) -> &str {
-        &*self
+        self
     }
 }
 
@@ -193,7 +210,7 @@ impl ops::Deref for ByteString {
 impl borrow::Borrow<str> for ByteString {
     #[inline]
     fn borrow(&self) -> &str {
-        &*self
+        self
     }
 }