+++ ntex-tokio/src/io.rs
@@ -93,7 +93,7 @@ impl Future for ReadTask {
                             Poll::Ready(Err(err)) => {
                                 log::trace!("read task failed on io {:?}", err);
                                 drop(io);
-                                let _ = this.state.release_read_buf(buf, new_bytes);
+                                this.state.release_read_buf(buf, new_bytes);
                                 this.state.close(Some(err));
                                 return Poll::Ready(());
                             }
@@ -540,7 +540,7 @@ mod unixstream {
                                 Poll::Ready(Err(err)) => {
                                     log::trace!("read task failed on io {:?}", err);
                                     drop(io);
-                                    let _ = this.state.release_read_buf(buf, new_bytes);
+                                    this.state.release_read_buf(buf, new_bytes);
                                     this.state.close(Some(err));
                                     return Poll::Ready(());
                                 }