+++ ntex-tokio/src/signals.rs
@@ -11,7 +11,7 @@ thread_local! {
 }
 
 /// Different types of process signals
-#[derive(PartialEq, Clone, Copy, Debug)]
+#[derive(PartialEq, Eq, Clone, Copy, Debug)]
 pub enum Signal {
     /// SIGHUP
     Hup,