+++ ntex-service/CHANGES.md
@@ -1,6 +1,12 @@
 # Changes
 
-## [0.3.2] - 2022-02-10
+## [0.3.3] - 2022-07-08
+
+* Revert cleanups
+
+## [0.3.2] - 2022-07-07
+
+* Add ?Sized to Rc service #125
 
 * Make AndThenFactory::new() public